TWAS-package {TWAS}	R Documentation

TWAS: A package for transcriptome-wide association studies

Description

The TWAS package provides a single function, twas, for jointly modeling two independent data sets: trait data (genome-wide association, GWA) and expression data (transcriptome-wide association, TWA). The analytical goal is to leverage the transcriptome data in analysis of the GWA data to inform the mechanisms underlying gene-trait associations.

[Package TWAS version 0.1.0 Index]