twas.gene.markers {TWAS}	R Documentation

Sample sets of markers proximal to genes

Description

twas.gene.markers is a list of character vectors of marker names for four sample gene names. The gene symbol is the name of each character vector.

Usage

twas.gene.markers

Format

An object of class list of length 4.

[Package TWAS version 0.1.0 Index]